Remove useless transformers & Remove transformers' children classes
Fix #117
Fix #118
