+++ src/main/scala/io/arlas/data/transform/ArlasTransformer.scala
@@ -70,27 +70,6 @@ abstract class ArlasTransformer(val dataModel: DataModel,
     Identifiable.randomUID(this.getClass.getSimpleName)
   }
 
-  /**
-    * List of datatype with matching scala types on https://stackoverflow.com/a/32899979/1943432
-    * ArrayType, MapType not supported because they need typing;
-    * DecimalType and StructType not supported because it is a DataType but AbstractDataType with different hierarch
-    */
-  protected def findDataTypeForValue(value: Any): DataType =
-    value match {
-      case b: Byte               => ByteType
-      case s: Short              => ShortType
-      case i: Int                => IntegerType
-      case l: Long               => LongType
-      case f: Float              => FloatType
-      case d: Double             => DoubleType
-      case s: String             => StringType
-      case b: Array[Byte]        => BinaryType
-      case b: Boolean            => BooleanType
-      case t: java.sql.Timestamp => TimestampType
-      case d: java.sql.Date      => DateType
-      case _                     => throw new DataFrameException(s"Unsupported type ${value.getClass.getName}")
-    }
-
 }
 
 case class DataFrameException(message: String) extends Exception(message)
