--- src/main/scala/io/arlas/data/transform/HmmProcessorBusiness.scala
@@ -1,44 +1,0 @@
-/*
- * Licensed to Gisaïa under one or more contributor
- * license agreements. See the NOTICE.txt file distributed with
- * this work for additional information regarding copyright
- * ownership. Gisaïa licenses this file to you under
- * the Apache License, Version 2.0 (the "License"); you may
- * not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *  http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing,
- * software distributed under the License is distributed on an
- * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
- * KIND, either express or implied.  See the License for the
- * specific language governing permissions and limitations
- * under the License.
- */
-
-package io.arlas.data.transform
-
-import io.arlas.data.model.{DataModel, MotionConfiguration, TempoConfiguration}
-import io.arlas.data.transform.ArlasTransformerColumns._
-import org.apache.spark.sql.SparkSession
-
-class WithArlasMovingState(dataModel: DataModel,
-                           spark: SparkSession,
-                           motionConfig: MotionConfiguration)
-    extends HmmProcessor(dataModel,
-                         spark,
-                         dataModel.speedColumn,
-                         motionConfig.movingStateModel,
-                         dataModel.idColumn,
-                         arlasMovingStateColumn,
-                         motionConfig.movingStateHmmWindowSize)
-
-class WithArlasTempo(dataModel: DataModel, spark: SparkSession, tempoConfig: TempoConfiguration)
-    extends HmmProcessor(dataModel,
-                         spark,
-                         arlasTrackDuration,
-                         tempoConfig.tempoModel,
-                         dataModel.idColumn,
-                         arlasTempoColumn,
-                         tempoConfig.tempoHmmWindowSize)
