--- src/test/scala/io/arlas/data/transform/WithArlasGeopointTest.scala
@@ -1,44 +1,0 @@
-/*
- * Licensed to Gisaïa under one or more contributor
- * license agreements. See the NOTICE.txt file distributed with
- * this work for additional information regarding copyright
- * ownership. Gisaïa licenses this file to you under
- * the Apache License, Version 2.0 (the "License"); you may
- * not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *  http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing,
- * software distributed under the License is distributed on an
- * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
- * KIND, either express or implied.  See the License for the
- * specific language governing permissions and limitations
- * under the License.
- */
-
-package io.arlas.data.transform
-
-import org.apache.spark.sql.DataFrame
-import io.arlas.data.sql._
-import io.arlas.data.transform.ArlasTransformerColumns._
-
-class WithArlasGeopointTest extends ArlasTest {
-
-  "WithArlasGeopoint" should "compute a geopoint from lat/lon" in {
-
-    val sourceDF = cleanedDF
-
-    val transformedDF: DataFrame = sourceDF
-      .enrichWithArlas(new WithArlasGeopoint(dataModel))
-
-    transformedDF
-      .collect()
-      .foreach(row => {
-        assert(
-          row.getAs[String](arlasGeoPointColumn) == row.getAs[Double]("lat") + "," + row
-            .getAs[Double]("lon"))
-      })
-  }
-
-}
